Corrected webserver authentication
Authentication now required as a first step when starting dialogue. User
is presented with a link to finish auth, then needs to enter code. Flow object is created, used to
generate link and fetch token, then deleted. Creds are sent to Policy
component in config. Added instrunctions on the page and changed
behaviour of the button as necessary.